xaprw001: create Mender artifact in post_image.sh
It was actually "wrong" to create the Mender artifact in the
post-build step, because at this point it has not yet generated
the new file-system images and we would have created the Mender
artifact using the file-system image from previous build.

Instead add an post-image script and create the Mender artifact
there, this will make sure that all file-system images are created
before we create the artifact.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>