Added support for openssl v1.1.0h to buildroot
Since a lot of software in the standard buildroot environment
currently relies on openssl v1.0.2o, and the API has changed
between both versions, it is not practical to upgrade openssl.

As an alternative, openssl v1.1.0h will be installed onto the
target in /opt/openssl11, so software that specifically needs
a newer openssl version (like enftun) can find it, but the
remainder of the target can continue using the older, default,
openssl library.

Based on https://patchwork.ozlabs.org/patch/910440/