Added Ethernet Gadget (with CDC Ethernet support) support
This results in the creation of the kernel module g_ether.
Inserting this module on the router board, results in the creation
of a L2 network interface usb0 on both sides.

This does not require any additional software on the host, and might
even work on Windows. I did not check this.