linux: move brcmfmac patches into global patches dir
The brcmfmac patches are not really specific the xaprw board.  At a
minimum, they should be applied to any kernel that builds the brcmfmac
driver.  It's safe to apply them to all kernels (the resulting source
files just won't be built), so it is easiest for us to manage them in
the global patches directory.

Most of the patches are upstream anyway, so we we upgrade to a new LTS
kernel, these patches can be dropped.