xaprc-common: place dynamic enftun credentials in data partition
enftun uses the TPM to authenticate to the ENF on its first use.  As
part of this initial handshake, it generates and stores a
public/private key pair to use for subsequent authentication attempts.
This key pair should be stored on the data partition so that it
persists across firmware updates.

This is purely a performance optimization.  The device may use the TPM
to authenticate each time, but doing so is inefficient.