amcl: fix sha256 hash of tarball
The AMCL folks deleted their github repo, so we had to move to hosting
our own forks.  The hash of the source tarball from Github must have
changed as part of this transition. I updated the Github source URL,
but not the tarball hash.  This patch updates the hash as well.